vuepress/project/browser-automation/requirements.txt
playwright==1.48.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2